Moved the Tcl layer into a separate file (crimp_tcl.tcl). Reworked the main file to source this file, immediately, and declared for when building a -pkg.
Reworked the Tcl layer itself, provide aggregate functions with
automatic selection of the proper C level primitive command, based on
image type, where possible (inver_* ==> invert).

Where decisions on type are not automatic the primitive is made a sub
method of a higher method (read_tk => read 2tk, etc.)

Combinations of this are possible as well, i.e. partially sub-methods,
partially automatic (convert, write).

Updated the akdemo to use the new Tcl layer, this simplifies quite a
bit as the typeof checking is now hidden in the new commands.