Moved the image type initialization out of find into an exported function, and changed the critcl setup to call it at the proper place in the package initialization sequence.
Extended the image type structure with information about the number of
'color' channels in an image of that type, and the names of these
channels. Added accessor returning the names of the channels in an
image as a list, and implicitly their number as well, in the length of
said list.

Renamed type field in image structure and updated all users. Had some
weird clash in gdb, preventing the latter from working with the field.

Removed some older macros from the main .tcl, they are in util.h
already.