Fixed solarize demo (boundary conditions).
Gave up on the Tcl layer reorg for the moment. Neither of the two
alternatives in the leaves [a20a604b8e] and [5b8699e126] feel properly
baked yet. Easier to keep the current hack and go forward with more
functionality I can later use to drive generalizations. Part of the
problem seems to be the implicit wish to auto-generate the higher API
from the low-level primitives.