<verbatim>
Created a new demo application which is properly modular.

I.e. this application pulls in the demo definitions from a directory,
automatically extending the GUI (toolbar), and also allowing each demo
to extend the GUI in a limited way.

Started to port the demos in akdemo over to this system.