<verbatim> Reworked the plot widget.
Added the ability to have a free-wheeling y-axis, i.e. auto-adapting
to the contents of the shown series.

Trouble with reconfiguring the axis after the creation of the plot
made me change things, i.e. now the plot is completely recreated on
every change of the series. This also fixes the problems I had with
keeping old data in the display.

Last change, the update of the display after changes is now properly
defered until the GUI is idle.