<verbatim> Extended the histogram demo.
(1) Compute cumulative density functions for RGB, luma, and SV of HSV.
(2) Show the various CDF in plots.
(3) Added histogram equalization (stretching) in RGB and HSV spaces (SV actually).
    Playing around shows that it might be sensible to have only S or V
    stretching in HSV.

The equalization in RGB shows how this naive operation (i.e. the
separate handling of the color channels) causes shifts in hue.
Equalization of the V generally makes the picture darker.

Oh, right, I seem to remember reading that for this stuff we have to
take out the gamma first, then equalize, and at last put the gamma
back in. I might try that as another demo.


Side problems with the plots, talk to Arjen:

	No control over the size of plotcharts. The histogram demo has
	so many plots now that the toplevel doesn't fit completely on
	the screen. Either the plots get smaller, or I (have to) split
	the demo by color space.

	After equalization we seem to have lots of zero counts around
	for intermediate values. The nature of the xyplot (linear
	lines between values) causes a moire which repeats the main
	plot at about half height. That looks ugly. Might have to use
	a different plot type to get rid of this. Barchart should
	work, except that it fills the area under the curve, which I
	don't want.