Demos.
The demos are now handled like they are in a matrix, with the choice
of images and actions as the two axes.

Choose an action and it applies to the current image.
Choose an image and the current action is applied.

So, choose an image and walk the toolbar to see the various actions.
Or, choose an action, maybe parameterize it, and then walk the images.