Convolution.
Simplified the primitive, ripped out the special case code handling of
image borders. This means that to avoid accessing pixels outside of
the input image we scan only the inner pixels of the input, and
consequently get a result shrunken by the kernel dimensions minus one.

To compensate for that we use the expand primitives implemented in the
last revisions to make the input appropriately larger before applying
the convolution.

The convolve method now accepts an option -border enabling the caller
to specify what type of border to use, with the default zero const
border. This shares the border type processing with 'expand',
therefore the code for that was moved into its own internal helper
command.

The method was further extended to accept more than one kernel, and
apply them in sequence.

The methods to make and transpose kernels have been reorganized into
an ensemble.