Undoing [0ac4353750], dropping the current 'resize' implementation.
Even with the gcd to reduce sample factors it is to easy to run into
large upsampling factors and massive memory consumption by the
transient intermediate images. Just need new/old width/height which
are relative prime (gcd == 1).

We need something which has all the three phases (upsample, convolve,
downsample) integrated, avoiding intermediate full images.

Starting to look towards the 'matrix' primitive and command now. More
general (projective transform), and having the resampling integrated
into the output loop. It needs some fixing tough, as is it is not
usable for the intended purpose.

When that is working however we can make resize and related commands
simple wrappers around it, for convenience.