Renamed the 'rangen' operator for clarity, and updated all users. (Note: 'rangen' can not only be read as 'ran-gen' (intended), but also as 'range-n', which is quickly confusing).
Swapped this operator's arguments as well. When width and height are
used together they are pretty much everywhere used in this order (w,
h). (Like in the result of crimp::dimensions). This enabled the
simplification of the calls (single 'dimensions' instead of separate
'width' and 'height').

Exposed the operator as 'crimp noise random'; updated the
documentation, and added demos.