I am impatient.
Having worked through and vetted most of the code in the parental
branch I want this code in the main branches, for regular use.

But there is still unvetted code, namely the 'filter::cleanup' and the
higher-order image registrations (rotation, scale, affine,
find-object), where I have run into roadblocks of understanding
myself, and where I defer to Kevin for vetting.

To get the vetted pieces into the mainline, without having the
unvetted parts this commit is made. It removes the unvetted pieces,
puts them into a new branch, and this new branch will then be
integrated with "works-for-critcl-2.1" and then "works-for-critcl-3".

The code is not lost, of course. It is still in the "x-smsaurabh" branch, and
when Kevin has worked through it, it can then be integrated also.