New branch.
Embed images into the infinite 2d plane, i.e. make them bounded
rectangles with a location in that plane. We had a start on that with
the operators for projective warp, storing the origin point in the
user-meta data. This information is now in the C level structures.

This commit provides the basic change, plus new accessors for the
geometry data, and updates all users of said geometry data in C to use
these new accessors.

The semantics of the various operations which could make use of the
location information has not changed yet. This is for future commits
in this branch.