Changed the image/volume construction APIss around a bit. API taking location information are now at the core. The old non-location APIs became macros injection their default location (0,0).
Added set of typed constructor taking location information.
Updated the warp init code to make use of the new constructors.

Beginning (with float/float) to update the binop core codes to handle
mismatched and -aligned images, i.e. of different sizes and translated
to each other. Instead of using additional macros for the special
cases we reuse the existing one and rely on the compiler to simplify
this as much as possible (i.e. constants, irrelevant operations, etc.)